pyproject.toml
[project]
name = "gigachat-token"
version = "0.1.0"
description = "GigaChat API Access Token Releaser"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiobotocore>=2.25.2",
    "environs>=14.5.0",
    "fastapi[standard]>=0.122.0",
    "psycopg[binary,pool]>=3.2.13",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "tzdata>=2025.2",
]
